Simplify bpacking files
